backend/requirements.txt
fastapi
uvicorn
pillow
easyocr
python-multipart
opencv-python
pydantic-settings
python-dotenv